nofix: syntax error in wikisyntax
The wikicréole parser cannot parse the following form, albeit correct:

    <<div|
      <<div id="a>>b"|content>>
    >>

The problem is the value of the `id' attribute: the parser doesen't accept
it since it contains two angle brackets `>>' which it interprets as the
end of the extension.
The error triggered says `syntax error in wikisyntax'.

With ohow however, that situation became much more frequent. Indeed, instead
of being compiled separately and then combined, the template and each wiki
are *first* combined (s/<<content>>/wiki content/) and then compiled as being
a single big wiki file.
In Ocsigen's template there are some enclosing `<<div' that surrounds the wiki
content and some wikis include documentation for the bind operator `(>>=)'
which code snippet look like this:

    <<pre id="VAL(>>=)" ...

That obviously triggers the problem described above because of the template's
enclosing `div's.

This "fix" adds to `ohow' another option, `--template', in order to restore
the old `<<content>>' insertion with the mixed compilation.
Still it does not fixes the wikicreole.mll parser: anyone can make it fail
by replicating this situation (unintentionally or not) in some wiki's content.